tests: Adjust assertions about LD_LIBRARY_PATH and ld.so.conf
Instead of being overly specific, let's assert that what we want to be
true is true. We want to see /overrides/lib/MULTIARCH in either the
ld.so.conf or the LD_LIBRARY_PATH, and /overrides/lib/MULTIARCH/aliases
in the LD_LIBRARY_PATH specifically.

If /overrides is not a symbolic link, then we want to find exactly
/overrides/lib/MULTIARCH in those places. If it is a symbolic link, then
we might see either /overrides/lib/MULTIARCH or
/usr/lib/pressure-vessel/overrides/lib/MULTIARCH; which one of those
we get is an implementation detail, but it must exist.

Signed-off-by: Simon McVittie <smcv@collabora.com>