./run: Always use pv-wrap --copy-runtime
In production Steam Linux Runtime builds, we always copy the runtime
into ./var and modify the copy. This is the only thing that can possibly
work in a Flatpak environment, and also the only thing that works
correctly in practice with Ubuntu and older Debian, due to their use of
a /sbin/ldconfig shell script wrapping /sbin/ldconfig.real (see
steam-runtime-tools!358).

For production Steam Linux Runtime builds, this is implicitly activated
by the presence of usr-mtree.txt.gz in the runtime, but when testing a
new runtime build by unpacking a new Platform tarball, it's confusing to
suddenly get different behaviour. Let's always use --copy-runtime mode,
so our behaviour is consistent.

Signed-off-by: Simon McVittie <smcv@collabora.com>