bubblewrap: Avoid a -Wjump-misses-init false-positive
When building with -Wjump-misses-init as part of a larger project, gcc
reports that we jump past initialization of cover_proc_dirs. This is
technically true, but we only use this variable in the case where it's
initialized, so that's harmless.

However, we can avoid this altogether by making the array static and
constant, which allows it to be moved from initialized data to read-only
data.

Signed-off-by: Simon McVittie <smcv@collabora.com>
Forwarded: https://github.com/containers/bubblewrap/pull/447