pv-wrap: Use local copy of bwrap in preference to system copy
The system copy might be older and less featureful, or might be
unnecessarily setuid, so let's try to use a version that is under our
control. If it works, great. If not, we can try a system copy, which
we might have to use anyway (for example because it's setuid).

Signed-off-by: Simon McVittie <smcv@collabora.com>