check-va-api: Do not stop if one decoding profile fails
Previously if we were able to create a configuration for a specific
decoding profile we assumed that this profile was working correctly, and
if it was not we returned an error.

Instead now we always fallback to the next decoding profile, until we
find one that works or run out of options.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>