Import host's icons inside the container
When creating a container, if we don't expose the host's icons, we
will end up with the default fallback for both icons and cursor theme.

This means being inconsistent with the components that are running
outside the container. This may also be a usability/accessibility issue
because the user defined color and/or size of the cursor will not be
kept.

Fixes: https://github.com/ValveSoftware/steam-runtime/issues/311

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>