tests: Run lint checks directly
lint checks shouldn't be connecting to the D-Bus session bus, so we
don't need dbus-run-session here.

With recent versions of Meson, we can use its built-in TAP parsing
instead of needing prove(1). Otherwise, we can just treat the test as
having unstructured output and let the exit code tell us whether it
succeeded.

Signed-off-by: Simon McVittie <smcv@collabora.com>