Avoid misleadingly-named g_spawn_check_exit_status() if possible
g_spawn_check_exit_status() was renamed to g_spawn_check_wait_status()
in GLib 2.70 because its name was misleading: the first argument was
always a wait-status (the output of wait(), waitpid() or waitid(),
encoding either an exit status or a terminating signal) rather than an
exit status.

Use the modern name in our code. When building for scout or
continue to provide our own implementation; when building for anything
else older than GLib 2.70, provide a macro to wrap the old, misleading
name.

Signed-off-by: Simon McVittie <smcv@collabora.com>