pv-runtime: Store resolved library per-arch to avoid memory leaks
For Vulkan ICDs, Vulkan layers and EGL ICDs, we reuse the same
IcdDetails for more than one architecture.

When we resolve the absolute path of the library for the second and
subsequent architectures, we need to avoid overwriting the path for the
previous architecture without freeing it. It might be good to refactor
this later to have separate arrays of IcdDetails per architecture, but
for now the least intrusive way to do this seems to be to put it in a
separate array slot per architecture.

Signed-off-by: Simon McVittie <smcv@collabora.com>