pv-runtime: If we find /usr/local/share/drirc.d, mirror that in container
If the user's graphics stack was not built with --prefix=/usr, for
example a Mesa developer with a local installation of Mesa in
--prefix=/usr/local, then we cannot necessarily expect it to load
/usr/share/drirc.d in the container: it is more likely to have
/usr/local/share/drirc.d compiled into it. To make sure that gets loaded,
we need to mirror the path from the graphics provider inside the
container.

In fact we already did that, but only in a special case for Flatpak
(in which graphics drivers are typically below /usr/lib/MULTIARCH/GL),
so we can mostly just expand that special case to non-Flatpak
installations.

For Flatpak, we know that we always have a mutable sysroot, so the
TAKE_FROM_PROVIDER_FLAGS_IF_CONTAINER_COMPATIBLE flag would be a no-op;
but in non-Flatpak we could conceivably be in the older mode where we
build up a long bubblewrap command-line, so we now want to pass in that
flag, because being unable to share the data directory is better than
just failing completely.

Because something might conceivably be relying on us putting the data
directory below /usr/share, we continue to do that as well.

The indentation around the GHashTable in this commit is a bit weird, to
keep the diff reviewable. It will be tidied up in a subsequent commit.

Resolves: https://gitlab.steamos.cloud/steamrt/steam-runtime-tools/-/issues/77
Signed-off-by: Simon McVittie <smcv@collabora.com>