pv-runtime: Refactor pv_runtime_collect_lib_data
If we want to locate drirc.d relative to Mesa EGL ICDs, Vulkan ICDs and
DRI drivers, it will probably be easier if pv_runtime_collect_lib_data()
takes a path in the provider namespace, so factor out the part that
inspects a capsule-capture-libs symlink and figures out the corresponding
path in the provider namespace.

By returning whether the symlink target could be determined (i.e. whether
it's a symlink), we can also elide some redundant stat() checks from
the caller.

The indentation in pv_runtime_collect_lib_data() is a bit weird to
keep the diffstat reasonable; it will be corrected in the next commit.

Signed-off-by: Simon McVittie <smcv@collabora.com>