graphics: Consider the sysroot while searching for modules
Previously we were always using _srt_check_library_presence() to locate
the loader libraries.

However if we are trying to use a graphical stack that is not from the
host system, it's not guaranteed that we can dlopen() the libraries
because they may have unmet dependencies.

For this reason when we have a sysroot that is not "/", we use instead
capsule-capture-libs to search for the correct library, under the
expected sysroot.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>