pv-runtime: Handle root-relative paths when not copying runtime
Two recent branches independently changed bits of PvRuntime for better
correctness in corner cases, and ended up passing in relative paths,
intending them to be interpreted as relative to the top of the sysroot
(provider or final container, as appropriate). This worked OK for the
code path where we modify a mutable copy of the sysroot in-place, but
did not work for the code path where we build up a long bwrap
command-line.

Fixes: a23cb089 "runtime: Use correct ld.so from provider"
Fixes: 485269b6 "pv-runtime: Use _srt_get_path_after to remove provider prefix"
Signed-off-by: Simon McVittie <smcv@collabora.com>