pv-runtime: Normalize paths in data_in_provider to be root-relative
This avoids the possibility of spurious diagnostics where we warn about
having found two different paths because we represented one of them as
foo/bar/baz and the other as /foo/bar/baz.

In the diagnostic messages, put the leading "/" back, to reduce
confusion about what they're meant to be relative to.

Signed-off-by: Simon McVittie <smcv@collabora.com>