pv-runtime: Find drirc.d adjacent to any known Mesa component
Previously, we assumed that we could find the ${prefix} of any Mesa
installation by basing it on the location of libGLX_mesa.so.0, and in
practice this is usually good enough for end-user systems.

However, Mike Blumenkrantz pointed out on #77 that in principle there
is nothing to stop someone from building a version of Mesa that only
has EGL, or a version that only has Vulkan. drirc.d is loaded by
multiple Mesa components, including EGL ICDs, Vulkan ICDs, and the
lower-level DRI drivers used to implement Mesa's EGL and GLVND ICDs.
Look for drirc.d adjacent to all of those.

Signed-off-by: Simon McVittie <smcv@collabora.com>