pv-runtime: Correct sense of a comparison
If the library is found in a directory not matching /*/lib/MULTIARCH,
we would have wrongly guessed that ${exec_prefix} is the root directory
by deleting a path that *didn't* match the known patterns. Of course
what was intended here is to delete a path that *does* match the known
patterns.

Fixes: 485269b6 "pv-runtime: Use _srt_get_path_after to remove provider prefix"
Signed-off-by: Simon McVittie <smcv@collabora.com>