pv-runtime: Clarify when library symlinks can end up in a subdirectory
We can divide the graphics drivers into two categories.

For graphics drivers that are found in a special directory (DRI, VA-API
and VDPAU), we will always get an absolute path from SrtSystemInfo
(and in fact we even have assertions to make sure this is the case),
so the graphics driver will always be of type ICD_KIND_ABSOLUTE on
success or ICD_KIND_NONEXISTENT on failure.

For graphics drivers that are found by reference to a manifest file
(Vulkan and EGL), we can either be given an absolute path or a basename,
so the graphics driver will be of type ICD_KIND_ABSOLUTE or
ICD_KIND_SONAME on success or ICD_KIND_NONEXISTENT on failure.
For ICD_KIND_ABSOLUTE, the path is arbitrary, so we put the symlink
in /overrides in a subdirectory off the default library search path to
avoid collisions. For ICD_KIND_SONAME, we always want to put the symlink
in the default search path, otherwise searching by SONAME will not
find it.

Either way, the use_subdir_for_kind_soname parameter is redundant:
for DRI etc., its value is irrelevant because we never actually see
ICD_KIND_SONAME, and for EGL etc., we always want it to be FALSE.
Remove the parameter altogether, and add assertions to check our
assumptions instead.

Signed-off-by: Simon McVittie <smcv@collabora.com>