pv-wrap: Add an option to mount /sys read/write
This allows developers to access debugfs, event tracing or similar
facilities, if allowed by filesystem permissions.

This seems risky to support for production use, since it would
potentially let games do things that are not long-term supportable, so
gate it behind an option equivalent to `flatpak run --devel`. I've made
the option generic so that we can use it to gate other similar features
in future.

Note that programs inside the pressure-vessel container cannot run
setuid, setgid or setcap executables for technical reasons.

Signed-off-by: Simon McVittie <smcv@collabora.com>