When the multiarch is not known, don't define _SRT_MULTIARCH
When the multiarch tuple was not known, we used to set _SRT_MULTIARCH to
the empty string.
However if we want to use the preprocessor directives to change the
behavior based on whenever the multiarch tuple is a known value, it will
be way easier to simply check if _SRT_MULTIARCH has been defined or not.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>