capture-libs: CAPTURE_FLAG_EVEN_IF_OLDER shouldn't apply for dependencies
The flag `CAPTURE_FLAG_EVEN_IF_OLDER` is not supposed to be considered
pertinent if we are checking a library dependency.

For this reason when we evaluate if a library dependency is a glibc
sub-lib, we should skip checking if `CAPTURE_FLAG_EVEN_IF_OLDER` is set.

This fixes an issue when capturing libraries with the pattern `gl:`,
while using an host system with glibc <= 2.33 and a container with glibc
>= 2.34. In this situation we were wrongly choosing `libpthread.so.0`
from the provider and `libc.so.6` from the container.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>
Co-authored-by: Simon McVittie <smcv@collabora.com>