build: Vendor our own copy of wflinfo
This was the last remaining third-party executable used for our
diagnostic checks that needed its packaging to be forked to build a
multiarch-qualified version. If we avoid needing that, then we can use
dependencies from an ordinary Linux distribution, making development
outside a Steam Runtime environment easier.

The next step for this would be to switch helpers/check-gl.c to use
libwaffle to create windows and contexts so that it can support more
platforms than just X11/GLX, and incorporate the parts of wflinfo's
functionality that we need into check-gl, so we only need to run one
check program rather than two.

Signed-off-by: Simon McVittie <smcv@collabora.com>