graphics: Load ICDs and layers to match Vulkan-Loader >= 1.2.198
It was previously non-compliant with the XDG basedir spec, but this
has now been fixed. Catch up with how it now works:

* XDG_CONFIG_HOME is now searched (previously it was not)

* XDG_DATA_HOME and ~/.local/share are now higher-precedence than
  XDG_DATA_DIRS

The fact that we search both XDG_DATA_HOME and ~/.local/share, even
if set, actually turns out to be a mistake (I misread the reference
Vulkan-Loader code), but it makes pressure-vessel accidentally work
around a Steam bug, so keep it for now. We deduplicate the search paths
automatically, so in the common case where XDG_DATA_HOME is unset or
points to ~/.local/share, it's harmless to look at it twice.

Signed-off-by: Simon McVittie <smcv@collabora.com>