urlopen: Allow steam.pipe to be tried either before or after x-d-p
In situations where Steam is providing the whole UI, such as Steam Deck
and Big Picture mode, we might want to prioritize steam.pipe higher than
x-d-p, so that Steam has the opportunity to choose what the UX will be
(launching an external browser, vs. showing the web page in-process
using CEF and steamwebhelper).

Signed-off-by: Simon McVittie <smcv@collabora.com>