pressure-vessel: Treat libXNVCtrl as part of graphics stack
MangoHUD uses this to get performance information for older NVIDIA
hardware via the NV-CONTROL X extension.

This is not technically part of the NVIDIA driver (it's part of the
nvidia-settings open-source module), and there's no guarantee that
NVIDIA users will have it installed at all, but we're probably more
likely to get a version compatible with the user's GPU and drivers if
we prefer to take it from the same place as the actual driver.

Resolves: https://gitlab.steamos.cloud/steamrt/steam-runtime-tools/-/issues/82
Signed-off-by: Simon McVittie <smcv@collabora.com>