graphics: Respect VK_DRIVER_FILES, VK_ADD_DRIVER_FILES
These are included in Vulkan-Loader 1.3.207, with a semantic change
(ignoring VK_ADD_DRIVER_FILES if VK_DRIVER_FILES or VK_ICD_FILENAMES
are set) merged for 1.3.208 but not yet releasd.

VK_DRIVER_FILES is a direct replacement for VK_ICD_FILENAMES, and is
preferred over VK_ICD_FILENAMES.

VK_ADD_DRIVER_FILES is prepended to the default search path if neither
of the overrides was used.

Resolves: steamrt/tasks#95
Signed-off-by: Simon McVittie <smcv@collabora.com>