_v2-entry-point: Make sure ./var is not a symlink
Some users might be tempted to redirect ./var to another filesystem,
but this will make the runtime a lot slower, more space-consuming and
less reliable, because we rely on being able to make hard links
between the distributed files and the temporary container root.

The correct way to relocate the Steam Linux Runtime depot is to set up
a Steam library on a suitable filesystem (which for best results should
be a fully-featured Unix filesystem like ext4, btrfs or xfs, although
NTFS can be made to work with some difficulty), and then use Steam's
"Properties -> Local Files -> Move install folder..." to move the
Steam Linux Runtime into it.

Helps: https://github.com/ValveSoftware/steam-runtime/issues/495
Signed-off-by: Simon McVittie <smcv@collabora.com>