pressure-vessel: Don't copy extended attributes
If we fall back to copying between btrfs and non-btrfs filesystems,
then we can fail to copy the btrfs.compression xattr; or if we are
copying from a filesystem that supports user xattrs to a filesystem that
does not (such as tmpfs), then we obviously won't be able to copy those.

We happen to know that pressure-vessel containers don't care about
extended attributes, so it's not a problem if we can't copy them.

Helps: https://github.com/ValveSoftware/steam-runtime/issues/495
Signed-off-by: Simon McVittie <smcv@collabora.com>