portal-listener: Stop listening on socket when portal is exiting
We should stop taking new requests as soon as we have made the decision
that we are going to exit. For the session bus, the way we stop getting
incoming requests is by releasing our bus name, but the equivalent for
a custom GDBusServer is to stop listening on our AF_UNIX socket.

Signed-off-by: Simon McVittie <smcv@collabora.com>