pv-wrap: Handle a larger subset of bwrap command-line syntax
When we add support for transparent emulation such as FEX-Emu, we'll
need to carry out rewriting for bwrap arguments generated from the
code we borrow from Flatpak. This will be easier to do if we support
more of bwrap's command-line options.

As before, we classify them by how many arguments they have, and which
arguments are paths in the host system.

Signed-off-by: Simon McVittie <smcv@collabora.com>