pv-runtime: Consolidate all driver SONAMEs into one capture-libs call
For drivers that are located by their absolute or relative path
(which we normalize into an absolute path), bind_icds() has significant
complexity to deal with putting the driver in an appropriate subdirectory
and the possibility of filename collisions, but for drivers that are
located by their SONAME, we don't need that complexity: only one
instance of the library per architecture can be the first in search-path
order, and we will put that instance first in our search path.

We can reduce the number of capsule-capture-libs calls by including
all such drivers in the "main" capsule-capture-libs invocation,
which we were already using to deal with dependencies.

We can also reduce the number of individual patterns in that "main"
invocation by combining drivers with their dependencies: in this case
they are both going to end up in the ${libdir} anyway.

Signed-off-by: Simon McVittie <smcv@collabora.com>