pv-runtime: Start to split up bind_icds() loop
We can do a first pass through the drivers to classify them into SONAME,
ABSOLUTE or missing, and process the SONAMEs. If there are no ABSOLUTE
paths, we can stop here. Otherwise, we only need to create the
subdirectory to contain symlinks to ABSOLUTE drivers once, then we can
enter a loop that does the rest of the more complicated processing of
ABSOLUTE drivers.

Signed-off-by: Simon McVittie <smcv@collabora.com>