pv-runtime: Delay computation of final_path
We want to split this loop into the parts that happen before and after
pv_runtime_capture_libraries(), which makes it harder to carry state
across that boundary.

Previously, we checked whether final_path already existed, but this
is in fact unnecessary: we already checked for collisions, and if a
collision was found, we'll be using individual destination directories
for each driver, which are newly-created and therefore ought to be
empty.

Signed-off-by: Simon McVittie <smcv@collabora.com>