pv-wrap: Bypass FEX-Emu overlay for most top-level directories
The FEX-Emu "rootfs" is conceptually just a /usr (plus parts of /etc and
/var), just like the Steam Runtime. However, the rootfs versions that
are currently distributed contain some files in /etc, /run and so on.

When we open files with opendir() or access() or similar, FEX-Emu will
transparently substitute files from the "rootfs" if they exist. However,
when we're building up a list of bind-mounts for a root filesystem,
this is counterproductive: bwrap will do the equivalent of
`mount --bind /oldroot/run/foo /newroot/run/foo`, which is not going
to work if /foo is really /home/me/.fex-emu/RootFS/Ubuntu_21_10/run/foo.

For several categories of file, and in particular /run and other API
filesystems, what we really want to do is to ignore the "rootfs" and
take the files and directories from the real root. Because FEX-Emu's
rewriting is path-based rather than identity-based, we can do this by
looking up the paths relative to /proc/self/root.

Signed-off-by: Simon McVittie <smcv@collabora.com>