pv-wrap: Adjust paths set up by Flatpak code for FEX-Emu rootfs
The code we've borrowed from Flatpak to deal with things like X11 and
D-Bus can find files and directories in either the real root filesystem,
or the FEX-Emu rootfs. If we naively pass the paths it has discovered to
bwrap, we'll find that they don't necessarily exist: for example, the
minimal aarch64 system I'm using for initial testing of this code
doesn't have /var/cache/fontconfig, but the rootfs does. This makes
bwrap fail when it tries to bind-mount /oldroot/var/cache/fontconfig
onto /newroot/var/cache/fontconfig.

To avoid this failure mode, look up the paths in the rootfs, and if
necessary rewrite them. For example, we can bind-mount
/oldroot/$FEX_ROOTFS/var/cache/fontconfig onto
/newroot/var/cache/fontconfig, and that will work fine.

Signed-off-by: Simon McVittie <smcv@collabora.com>