pv-runtime: Be more compatible with interpreters like FEX-Emu
A common approach to transparent CPU-architecture emulation is to have
an overlay with binaries for the emulated architecture, and transparently
use those (as though they were an overlayfs over the real filesystem)
when an emulated process accesses their paths.

For example, FEX-Emu emulates x86 on aarch64 systems. When an aarch64
process opens an architecture-dependent file like /usr/lib64/libc.so.6
or /etc/ld.so.cache, it gets the real file; when an emulated x86 process
opens an architecture-dependent file, if that file exists in the
FEX-Emu "rootfs", FEX-Emu will transparently redirect the file access
to the copy in the "rootfs".

When we enter the Steam Runtime container, we need two environments
inside the container: an aarch64 environment that can run the FEX-Emu
interpreter itself, and an x86 environment that can run
pressure-vessel-adverb, Steam Runtime executables, and the actual game.

To arrange for this to happen, we can use a subset of the real aarch64
operating system as the root of the container, and set up a new FEX-Emu
rootfs in /run/pressure-vessel/interpreter-root. This means that
aarch64 executables in the Steam Runtime container see the real host OS,
but x86 executables see the Steam Runtime.

Signed-off-by: Simon McVittie <smcv@collabora.com>