_srt_get_helper: Don't use timeout(1) when running inside a Snap app
Canonical's Snap packaging of Steam has a restrictive AppArmor profile
that doesn't allow running timeout(1), resulting in all of our helper
subprocesses failing.

This is a short-term solution: it will result in Help -> System
Information waiting forever if one of our helper subprocesses deadlocks
or otherwise hangs (for example as a result of a faulty driver).

Signed-off-by: Simon McVittie <smcv@collabora.com>