runtime: Remove over-allocation of arrays keyed by multiarch index
Previously these were the same length as pv_multiarch_tuples, but they
don't actually need to be. The only reason pv_multiarch_tuples has an
extra element is so we can use it as a NULL-terminated array (like a
GStrv), but we don't do that with these arrays, which can be sparse
(because an ICD might exist on one architecture but not on another).

Signed-off-by: Simon McVittie <smcv@collabora.com>