glib-backports: Backport g_memdup2
This is a fixed version of g_memdup(), which is deprecated because it's
vulnerable to integer overflows if the byte_size is attacker-controlled.
I'm not actually intending to use it for anything that could be
attacker-controlled, but doing this backport seems clearer than proving
that our use of the deprecated function is safe.

Signed-off-by: Simon McVittie <smcv@collabora.com>