pv-runtime: Add a fast-path for drivers that are symlinked or hardlinked
In recent Mesa versions, all DRI drivers are hard-links to one of a
small number of large "megadrivers", typically one for Gallium drivers
and one for "classic" drivers.

When dealing with these drivers, there's little point in processing them
one by one: everything capsule-capture-libs does is dependent only on
the content of the driver, so whatever we would have done for
(for example) iris_dri.so (a Gallium driver), we'll do exactly the same
for the nouveau_dri.so that is another name for the same file.

Instead, process only the first representative of each file, and then
copy its result (whether positive or negative) for the others.

Signed-off-by: Simon McVittie <smcv@collabora.com>