_srt_x86_cpuid: Optionally skip capability checking
`__get_cpuid()` uses CPUID leaves 0 (highest function parameter and
manufacturer ID) and 0x8000_0000 (highest extended function parameter)
to check whether the requested leaf is available, but this doesn't work
for pseudo-CPUIDs coming from the hypervisor, such as the hypervisor
interface starting at 0x4000_0000. For those, we will have to do the
capability checks ourselves. Add a parameter to indicate that we
already did so.

Signed-off-by: Simon McVittie <smcv@collabora.com>