Add basic detection of virtualization, hypervisors and emulation
At the moment this handles anything that systemd detects via the easier
mechanisms like CPUID or DMI ID, plus FEX-Emu. It does not handle more
complicated detection mechanisms such as PCI devices or DeviceTree.

As a design-for-test feature, the low-level detection function can be
given mock information to use instead of the real system.

Signed-off-by: Simon McVittie <smcv@collabora.com>