Reduce code duplication with a new load_icd_from_json function
Group together the EGL and Vulkan functions that load an ICD from a JSON
metadata.

This reduces the code duplications and allows us to expand the number of
elements that can be parsed without increasing the `load_json()` function
parameters (e.g. adding support for the new `is_portability_driver`).

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>