Support the new Vulkan portability driver flag
Vulkan ICD manifests recently added a new file format 1.0.1 version.

This new version introduces a "is_portability_driver" flag that is used
to signal drivers that are not entirely compliant with the Vulkan API.

With this commit we add support in s-r-t to parse this new flag and
introduce a new loadable issue called "SRT_LOADABLE_ISSUES_API_SUBSET",
to make it easier to spot ICDs that are not fully API compliant.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>