scout-on-soldier: Add a convenience script for developers
It's convenient to be able to use a command like

    .../SteamLinuxRuntime_soldier/run -- xterm

to enter the soldier environment for some quick testing, but there isn't
a direct equivalent for scout-on-soldier. Add one.

We have to make an assumption about the location of
SteamLinuxRuntime_soldier, because officially only Steam knows where to
find each compatibility tool; but we can guess that most developers will
have put SteamLinuxRuntime_soldier and SteamLinuxRuntime in the same
Steam library, whether that's the default or not.

Signed-off-by: Simon McVittie <smcv@collabora.com>