pv-runtime: Really capture dependencies of captured drivers
If we call fstatat() and let it follow symlinks, then we'll always think
the symlink was not created. In the common case where it is pointing
to a driver that doesn't exist in the current namespace (but will exist
in the container), fstatat() will fail with ENOENT; or if the driver
also exists in the current namespace, then S_ISLNK will return false.

Fixes: 89e7738 "pv-runtime: Do more fd-relative I/O in bind_icds()"
Resolves: steam-runtime-tools#91
Signed-off-by: Simon McVittie <smcv@collabora.com>