populate-depot: Calculate mtree from the extracted runtimes
If we use the compressed runtime to generate the mtree, we need to
extract every single file, on the fly, to calculate their hash.
But given that when we call `write_lookaside()` we already have entirely
extracted the archive on disk, we can use that uncompressed directory to
skip re-extracting the files a second time.

Additionally this will allow us to remove some files from the runtimes
that can be safely omitted to save space, e.g. documentation or support
for additional languages.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>