pv-runtime: Create /usr/share/defaults/etc/ld.so.conf, for Solus
Solus' ldconfig is hard-coded to read this path as its default, rather
than the /etc/ld.so.conf used in all other known distributions, in an
effort to be able to run with an empty /etc. When Steam's LD_LIBRARY_PATH
runtime runs inside a pressure-vessel container, it relies on ldconfig
picking up the configuration that was set up by pressure-vessel, so we
need this path to be a symlink to the location we have set up.

Because this path is below /usr, it'll only work for the mutable-sysroot
code path, but in practice that's the one that Steam uses in production
anyway, so this limitation seems reasonable.

Resolves: https://github.com/ValveSoftware/steam-runtime/issues/510
Signed-off-by: Simon McVittie <smcv@collabora.com>