pv-runtime: Start generalizing creation of ld.so.cache symlinks
This is mostly just refactoring, but it introduces the concept that
creating the symlink might fail. This can't actually happen in practice,
because the only failure mode that is implemented right now is bwrap
being NULL, which cannot really happen in these call sites.

These alternate filenames won't be critical on most OS configurations,
so if creating the symlink fails, we just warn and try to continue.

The slightly odd error exit path will get more use in subsequent commits.

Signed-off-by: Simon McVittie <smcv@collabora.com>