pv-runtime: Create symlinks in /usr by modifying mutable sysroot
By the time we're running bubblewrap, it's too late to modify /usr:
we're going to mount it read-only, so writes will fail. However, in the
configuration with a mutable sysroot (which is what we actually do in
production), we can modify the mutable sysroot before that point
instead.

This code isn't used yet, but will be used in a subsequent commit.

Signed-off-by: Simon McVittie <smcv@collabora.com>