Add basic steam-runtime-launcher-service integration
If run with STEAM_COMPAT_LAUNCHER_SERVICE=container-runtime, we'll
inject commands at the container layer and they'll run in a "pure"
soldier/sniper environment. This is appropriate for native soldier/sniper
games that run with no further compatibility layer, as in
ValveSoftware/steam-runtime#508.

If run with STEAM_COMPAT_LAUNCHER_SERVICE=scout-in-container, we'll
inject commands inside the layered scout-on-soldier runtime (or
potentially scout-on-sniper, if we pivot to that at some point) and
they'll run in a scout-compatible environment.

Signed-off-by: Simon McVittie <smcv@collabora.com>